requirements.txt
boto3==1.34.34
langchain~=0.3.27
langchain-openai~=0.3.31
python-dotenv==1.0.1
tabulate==0.9.0
setuptools~=80.9.0
pydantic~=2.11.7